requirements.txt
pymongo[srv]==4.6.1
python-dotenv==1.0.0
pandas==2.1.4
dnspython==2.4.2
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0